requirements.txt
# requirements.txt
# Energy Fuse Monitoring & Archiving Pipeline
# Home Assistant + InfluxDB 1.8 → MariaDB
# TEK5370 Smart Grid & IoT – 2025

# === Core dependencies ===
python-dotenv>=1.0.1              # Load .env files automatically
influxdb-client>=1.40.0           # Official InfluxDB 2.x client
influxdb==5.3.2                   # InfluxDB 1.x client (you are using InfluxDB 1.8)
pandas>=2.0.0                     # Data handling & export
SQLAlchemy>=2.0.0                 # Database engine
PyMySQL>=1.1.0                    # MariaDB/MySQL driver
urllib3==1.26.7                   # Use older version to support OpenSSL

# === Data science & visualization (optional but useful) ===
numpy>=1.24.0
seaborn>=0.13.0
matplotlib>=3.7.0
plotly>=5.18.0
scikit-learn>=1.3.0
xgboost>=2.0.0